Modified the Addition and Subtraction class. Added proper code for each method.
Modified the MapCreatorFormTxt class. Added proper code for the
scanTerrain method so that it can read from the txt file. Also set up
the getter and the setter methods

Added proper code to the Literal class for the getter method that
returns Integer Object.

BinaryExpression getter for Integer Object was modified to return that
Object.